Added alert dialog

+++ app/src/main/java/com/example/todolist/DashboardActivity.kt
@@ -109,8 +109,17 @@ class DashboardActivity : AppCompatActivity() {
                             activity.updateToDo(list[p1])
                         }
                         R.id.menu_delete->{
-                            activity.dbHandler.deleteToDo(list[p1].id)
-                            activity.refreshList()
+                            val dialog = AlertDialog.Builder(activity)
+                            dialog.setTitle("Are you sure?")
+                            dialog.setMessage("Do you want to delete this task?")
+                            dialog.setPositiveButton("Yes") { _: DialogInterface, _: Int ->
+                                activity.dbHandler.deleteToDo(list[p1].id)
+                                activity.refreshList()
+                            }
+                            dialog.setNegativeButton("No") { _: DialogInterface, _: Int ->
+                            }
+                            dialog.show()
+
                         }
                         R.id.menu_mark_as_completed->{
                             activity.dbHandler.updateToDoItemCompletedStatus(list[p1].id,true)
